fixes 1 module5-task1
